Add react-native SparkBarChart
