+++ package.json
@@ -131,6 +131,7 @@
     "react-dom": "^17.0.2",
     "react-native-svg": "^12.1.1",
     "react-native-web": "0.17.1",
+    "react-spring": "^9.4.2",
     "rimraf": "^3.0.0",
     "sass-loader": "^10.1.1",
     "semver": "^6.3.0",
