+++ packages/polaris-viz-core/src/components/SparkBarSeries/components/Bar/index.ts
