+++ packages/polaris-viz-core/src/components/SparkBarSeries/components/Bar/tests/Bar.test.tsx
