+++ packages/polaris-viz-core/src/components/SparkBarSeries/components/index.ts
