+++ packages/polaris-viz-core/src/components/SparkBarSeries/components/tests/Bar.test.tsx
