+++ packages/polaris-viz-core/src/components/SparkBarSeries/index.ts
@@ -1,0 +1,1 @@
+export {SparkBarSeries} from './SparkBarSeries';
