+++ packages/polaris-viz-core/src/utilities/get-animation-trail.ts
