+++ packages/polaris-viz-native/src/components/SparkBarChart/index.ts
@@ -1,0 +1,1 @@
+export {SparkBarChart} from './SparkBarChart';
