+++ packages/polaris-viz/src/components/SparkBarChart/SparkBarChart.scss
@@ -3,7 +3,3 @@
 .VisuallyHidden {
   @include visually-hidden;
 }
-
-.ComparisonLine {
-  stroke-linecap: round;
-}
