+++ packages/polaris-viz/src/utilities/index.ts
@@ -6,7 +6,6 @@ export {RightAngleTriangle} from './RightAngleTriangle';
 export {getBarXAxisDetails} from './get-bar-xaxis-details';
 export {getLongestLabelDetails} from './get-longest-label-details';
 export {getMaxDiagonalDetails} from './get-max-diagonal-details';
-export {getAnimationTrail} from './get-animation-trail';
 export {getPathLength} from './get-path-length';
 export {getPointAtLength} from './get-point-at-length';
 export {normalizeData} from './normalize-data';
